refactor: sidemenu and navigation items component
